1. Removed unneeded includes for <new>. 2. Fixed up namespace use (this churned nearly everything). 3. Re-enabled Java and Ficl to completely test. I/O unit tests all work. Haven't done the rest yet.
git-svn-id: svn://graphite/desperado/trunk@142 3852cbcf-8179-446f-bc45-53e98ba5d360

+++ Desperado/BufferInput.cpp
@@ -57,7 +57,6 @@
  */
 
 
-#include <new>
 #include "com/diag/desperado/errno.h"
 #include "com/diag/desperado/target.h"
 #include "com/diag/desperado/string.h"
@@ -191,7 +190,7 @@ ssize_t BufferInput::operator() (
         rc = 0;
     } else {
     	rc = this->size - this->offset;
-        if (rc > maximum) {
+        if (rc > static_cast<ssize_t>(maximum)) {
         	rc = maximum;
         }
         memcpy(bp, &(this->buffer[this->offset]), rc);
